Add new plotting functionalities for annual quantities, land use, per capita metrics, and self-sufficiency analysis.
